refactor: transform the exam card list generator into a single component
