+++ utils/convertToISODate.ts
@@ -1,0 +1,7 @@
+export default function convertToISODate(eventDate: string) {
+    const [datePart, timePart] = eventDate.split(' ')
+    const [day, month, year] = datePart.split('/').map(Number)
+    const [hours, minutes] = timePart.split(':').map(Number)
+
+    return new Date(year, month - 1, day, hours, minutes)
+}
